Fixed typos in docs, added new docs for cltypes, updated types
